Convert Aircraft Manuals app to align with Charts app. Refactor ChartsApp into common base class and a small specialisation. Change PlaneManualApp to derive from the new DocumentsApp base class.
