+++ src/avitab/apps/CMakeLists.txt
@@ -4,6 +4,7 @@ target_sources(avitab_common PRIVATE
     ${CMAKE_CURRENT_LIST_DIR}/HeaderApp.cpp
     ${CMAKE_CURRENT_LIST_DIR}/AppLauncher.cpp
     ${CMAKE_CURRENT_LIST_DIR}/ChartsApp.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/DocumentsApp.cpp
     ${CMAKE_CURRENT_LIST_DIR}/NotesApp.cpp
     ${CMAKE_CURRENT_LIST_DIR}/About.cpp
     ${CMAKE_CURRENT_LIST_DIR}/PlaneManualApp.cpp
