+++ src/avitab/apps/ChartsApp.h
@@ -18,96 +18,13 @@
 #ifndef SRC_AVITAB_APPS_CHARTSAPP_H_
 #define SRC_AVITAB_APPS_CHARTSAPP_H_
 
-#include <memory>
-#include <vector>
-#include "src/platform/Platform.h"
-#include "App.h"
-#include "src/gui_toolkit/widgets/TabGroup.h"
-#include "src/gui_toolkit/widgets/Page.h"
-#include "src/gui_toolkit/widgets/Window.h"
-#include "src/gui_toolkit/widgets/List.h"
-#include "src/gui_toolkit/widgets/PixMap.h"
-#include "src/gui_toolkit/widgets/Checkbox.h"
-#include "src/gui_toolkit/widgets/Container.h"
-#include "src/gui_toolkit/widgets/Label.h"
-#include "src/gui_toolkit/Timer.h"
-#include "src/libimg/Image.h"
-#include "src/libimg/stitcher/Stitcher.h"
-#include "src/maps/OverlayedMap.h"
-#include "src/maps/sources/PDFSource.h"
-#include "components/FilesysBrowser.h"
+#include "DocumentsApp.h"
 
 namespace avitab {
 
-class ChartsApp: public App {
+class ChartsApp: public DocumentsApp {
 public:
     ChartsApp(FuncsPtr appFuncs);
-    void onMouseWheel(int dir, int x, int y) override;
-private:
-    Timer updateTimer;
-
-    void resetLayout();
-
-    std::shared_ptr<TabGroup> tabs;
-
-    std::shared_ptr<Page> browsePage;
-    std::shared_ptr<Window> browseWindow;
-    std::shared_ptr<List> list;
-    FilesystemBrowser fsBrowser;
-    std::vector<platform::DirEntry> currentEntries;
-    std::shared_ptr<maps::OverlayConfig> overlays;
-
-    void createBrowseTab();
-    void showDirectory();
-    void setFilterRegex(const std::string regex);
-    void showCurrentEntries();
-    void upOneDirectory();
-    void onSettingsToggle(bool forceClose = false);
-    void onDown();
-    void onUp();
-    void onSelect(int data);
-
-    struct PdfPage {
-        std::string path;
-        std::shared_ptr<Page> page;
-        std::shared_ptr<Window> window;
-        std::shared_ptr<PixMap> pixMap;
-        std::shared_ptr<img::Image> rasterImage;
-        std::shared_ptr<maps::PDFSource> source;
-        std::shared_ptr<img::Stitcher> stitcher;
-        std::shared_ptr<maps::OverlayedMap> map;
-        int panStartX = 0, panStartY = 0;
-    };
-
-    using PageInfo = std::shared_ptr<PdfPage>;
-    std::vector<PageInfo> pages;
-
-    void createPdfTab(const std::string &pdfPath);
-    void removeTab(std::shared_ptr<Page> page);
-    void setupCallbacks(PageInfo tab);
-    void loadFile(PageInfo tab, const std::string &pdfPath);
-    void setTitle(PageInfo tab);
-    PageInfo getActivePdfPage();
-    void onNextPage();
-    void onPrevPage();
-    void onPlus();
-    void onMinus();
-    void onScrollUp();
-    void onScrollDown();
-    void onRotate();
-    void onPan(int x, int y, bool start, bool end);
-    bool onTimer();
-
-    Settings::PdfReadingConfig  settings;
-    std::shared_ptr<Container> settingsContainer;
-    std::shared_ptr<Label> settingsLabel;
-    std::shared_ptr<Checkbox> mouseWheelScrollsCheckbox;
-    void showAppSettings();
-
-    enum class VerticalPosition { Top, Centre, Bottom };
-    enum class HorizontalPosition { Left, Middle, Right };
-    enum class ZoomAdjust { None, Height, Width, All };
-    void positionPage(PageInfo tab, VerticalPosition vp, HorizontalPosition hp, ZoomAdjust za = ZoomAdjust::None);
 };
 
 } /* namespace avitab */
