+++ src/avitab/apps/PlaneManualApp.h
@@ -18,31 +18,15 @@
 #ifndef SRC_AVITAB_APPS_PLANEMANUALAPP_H_
 #define SRC_AVITAB_APPS_PLANEMANUALAPP_H_
 
-#include <vector>
-#include "App.h"
-#include "src/platform/Platform.h"
-#include "src/gui_toolkit/widgets/MessageBox.h"
-#include "components/FilesysBrowser.h"
+#include "DocumentsApp.h"
 
 namespace avitab {
 
-class PlaneManualApp: public App {
+class PlaneManualApp: public DocumentsApp {
 public:
     PlaneManualApp(FuncsPtr appFuncs);
-    void show() override;
-    void onMouseWheel(int dir, int x, int y) override;
 private:
-    std::string currentAircraft;
-    std::shared_ptr<MessageBox> errorMsg;
-    std::string currentPath;
-    FilesystemBrowser fsBrowser;
-    std::shared_ptr<App> childApp;
-
-    void showFileSelect();
-    void onSelect(const std::vector<platform::DirEntry> &entries, size_t chosenIndex);
-    void onSelectionClosed();
-
-    void showAircraftFolder();
+    std::string findStartDirectory();
 };
 
 } /* namespace avitab */
